// frontend/src/types/generated/EventAnchor.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AutoName } from "./AutoName";

export type EventAnchor = { cnId: string; nameCn: string; cnStart: number; cnEnd: number; offsetSecs: number; nameEn: string | null; nameEnAuto: AutoName | null };