// frontend/src/types/generated/ReviewOutfit.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AutoName } from "./AutoName";
import type { SkinPrice } from "./SkinPrice";

export type ReviewOutfit = { cnGetTime: number; enGetTime: number | null; skinId: string; charId: string; skinName: string; charName: AutoName | null; portraitPath: string | null; colors: Array<string>; price: SkinPrice };